Fixed weird issue where headers were not extracted in the throttled task runner by doing it beforehand with final vars.
